Fix #13: Wii Balance Board
The implementation is now using the Device Calibration. However each sensor reports its data in absolute Gramms; the calibration data provides values for 0Kg, 17Kg and 34Kg. The absolute value of each sensor is interpolated between the 0Kg and 34 Kg (0Kg = 0; 17Kg = 127; 34Kg = 255). For each output axis, two sensor are joined (calibrated values are added and divided by 4 to get a value between 0 and 127) and then either added or subtractet from the Null State Value 127 to get a value range of 0 - 255.
This means if a person has a bodyweight less than 68KG, it is not able to achieve the full value range. So at one time the implementation should be changed to take the bodyweight into account and to provide relative values.